Fix Minimap
Fix scale of the minimap and map
Fix the problem of the minimap not load when enter in a new room
