--- Instance3/Assets/Map/GestionMap/Scripts/RoomManager1.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: fffe0c2d683b583418a8fba41753e8b6
