+++ Instance3/Assets/Map/MapUi/MapUitilePalette.prefab
@@ -71,8 +71,8 @@ Transform:
   m_GameObject: {fileID: 2491701876083388805}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 20.63, y: -0.497, z: 0}
-  m_LocalScale: {x: 15, y: 15, z: 1}
+  m_LocalPosition: {x: 24.6, y: -0.1, z: 0}
+  m_LocalScale: {x: 20.71875, y: 18.285938, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 7918264037109882946}
@@ -158,7 +158,7 @@ Transform:
   m_GameObject: {fileID: 3450977074169320603}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 76.09, y: -0.497, z: 0}
+  m_LocalPosition: {x: 106.9, y: 3.9, z: 0}
   m_LocalScale: {x: 15, y: 20, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -245,7 +245,7 @@ Transform:
   m_GameObject: {fileID: 4744451099296980170}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 56.44, y: -0.497, z: 0}
+  m_LocalPosition: {x: 83.8, y: 4.2, z: 0}
   m_LocalScale: {x: 20, y: 15, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -332,7 +332,7 @@ Transform:
   m_GameObject: {fileID: 5529929862489832371}
   serializedVersion: 2
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 93.54, y: -0.497, z: 0}
+  m_LocalPosition: {x: 122.8, y: 1.6, z: 0}
   m_LocalScale: {x: 2, y: 4, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -548,7 +548,7 @@ Transform:
   m_GameObject: {fileID: 7316673021161638852}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 36.8, y: 17.7, z: 0}
+  m_LocalPosition: {x: -16.5, y: 0.2, z: 0}
   m_LocalScale: {x: 20.75, y: 11.25, z: 2.075}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -635,7 +635,7 @@ Transform:
   m_GameObject: {fileID: 7551962722537016966}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -0.497, y: -0.497, z: 0}
+  m_LocalPosition: {x: 55.6, y: 4, z: 0}
   m_LocalScale: {x: 20, y: 20, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
