+++ Instance3/Assets/Map/MapUi/Scripts/GiveColliderForCinemachine.cs
@@ -1,5 +1,4 @@
 using System;
-using Unity.VisualScripting;
 using UnityEngine;
 
 public class GiveColliderForCinemachine : MonoBehaviour
